some fixes

+++ plugins/module_utils/virtual_disk.py
@@ -13,7 +13,7 @@ from ..module_utils.typed_classes import (
     TypedVirtualDiskToAnsible,
     TypedTaskTag,
 )
-from typing import Union, Dict, List, Any
+from typing import Dict, List, Any, Optional
 
 from .rest_client import RestClient
 from ..module_utils.utils import PayloadMapper
@@ -23,12 +23,12 @@ from ..module_utils import errors
 class VirtualDisk(PayloadMapper):
     def __init__(
         self,
-        name: Union[str, None] = None,
-        uuid: Union[str, None] = None,
-        block_size: Union[int, None] = None,
-        size: Union[int, None] = None,
+        name: Optional[str] = None,
+        uuid: Optional[str] = None,
+        block_size: Optional[int] = None,
+        size: Optional[int] = None,
         # allocated_size: int = None,
-        replication_factor: Union[int, None] = None,
+        replication_factor: Optional[int] = None,
     ):
         self.name = name
         self.uuid = uuid
@@ -59,10 +59,6 @@ class VirtualDisk(PayloadMapper):
 
     def to_hypercore(self) -> Dict[Any, Any]:
         raise NotImplementedError()
-        # return dict(
-        #     searchDomains=self.search_domains,
-        #     serverIPs=self.server_ips,
-        # )
 
     def to_ansible(self) -> TypedVirtualDiskToAnsible:
         return dict(
@@ -101,7 +97,7 @@ class VirtualDisk(PayloadMapper):
     @classmethod
     def get_by_name(
         cls, rest_client: RestClient, name: str, must_exist: bool = False
-    ) -> Union[VirtualDisk, None]:
+    ) -> Optional[VirtualDisk]:
         result = rest_client.list_records("/rest/v1/VirtualDisk", query=dict(name=name))
         if must_exist and not result:
             raise errors.ScaleComputingError(
@@ -141,7 +137,9 @@ class VirtualDisk(PayloadMapper):
             query=dict(filename=file_name, filesize=file_size),
         )
 
-    def send_delete_request(self, rest_client: RestClient) -> TypedTaskTag:
-        return rest_client.delete_record(
-            f"/rest/v1/VirtualDisk/{self.uuid}", check_mode=False
-        )
+    def send_delete_request(self, rest_client: RestClient) -> Optional[TypedTaskTag]:
+        if self.uuid:
+            return rest_client.delete_record(
+                f"/rest/v1/VirtualDisk/{self.uuid}", check_mode=False
+            )
+        return None
